Move main file to index.js
